pyproject.toml
# Python dependencies for CISC experiment
[project]
name = "cisc-experiment"
version = "0.1.0"
requires-python = ">=3.9"
dependencies = [
    "torch>=2.0.0",
    "transformers>=4.30.0",
    "datasets>=2.14.0",
    "hydra-core>=1.3.0",
    "wandb>=0.15.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "numpy>=1.24.0",
    "omegaconf>=2.3.0",
    "accelerate>=0.20.0",
]
